// src/app/open-api/generated/model/alias.ts
/**
 * GED API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Alias { 
    readonly id?: number;
    primaryWord: string;
    aliases: Array<string>;
}